+++ chrome_agent/manifest.json
@@ -14,6 +14,7 @@
     "scripts": [
       "js/lib/axiom_base.amd.concat.js",
       "js/lib/wash.amd.concat.js",
+      "js/chrome_agent_client.js",
       "js/chrome_command.js",
       "js/chrome_agent.js",
       "js/extension/background.js"
